refactor: consolidate watermarking-functions to root as canonical source
- Sync root watermarking-functions with extended detection statistics
- Remove duplicate from watermarking-docker (now copied during build)
- Add build.sh to watermarking-docker for consistent build process
- Update watermarking-api/build.sh to use root source

This enables extracting watermarking-functions to a separate private
repo while making the rest of the codebase public.
