+++ watermarking-docker/.gitignore
@@ -1,5 +1,8 @@
 .DS_Store
 
+# Build artifact (copied from root watermarking-functions/ during build)
+watermarking-functions/
+
 #From: https://github.com/github/gitignore/blob/master/Node.gitignore 
 
 # Logs
