+++ watermarking-functions/DETECTION.md
